+++ __tests__/client/AbstractClient.test.js
@@ -608,7 +608,8 @@ describe('Fix bugs', () => {
       .mock({
         name: 'generate_token',
         matcher: (url, opts) =>
-          url === 'https://oauth.me/' && opts.body._streams[7] === 'password',
+          url === 'https://oauth.me/' &&
+          opts.body.get('grant_type') === 'password',
         response: {
           body: {
             access_token: 'an_access_token',
@@ -622,9 +623,12 @@ describe('Fix bugs', () => {
       })
       .mock({
         name: 'refresh_token',
-        matcher: (url, opts) =>
-          url === 'https://oauth.me/' &&
-          opts.body._streams[1].match('refresh_token'),
+        matcher: (url, opts) => {
+          return (
+            url === 'https://oauth.me/' &&
+            opts.body?.get('refresh_token') === 'refresh_token'
+          );
+        },
         response: {
           body: {
             access_token: 'a_refreshed_token',
@@ -702,7 +706,8 @@ describe('Fix bugs', () => {
       .mock({
         name: 'generate_token',
         matcher: (url, opts) =>
-          url === 'https://oauth.me/' && opts.body._streams[7] === 'password',
+          url === 'https://oauth.me/' &&
+          opts.body.get('grant_type') === 'password',
         response: {
           body: {
             access_token: 'an_access_token',
@@ -718,7 +723,7 @@ describe('Fix bugs', () => {
         name: 'refresh_token',
         matcher: (url, opts) =>
           url === 'https://oauth.me/' &&
-          opts.body._streams[1].match('refresh_token'),
+          opts.body.get('refresh_token') === 'refresh_token',
         response: {
           body: {
             access_token: 'a_refreshed_token',
@@ -789,7 +794,8 @@ describe('Fix bugs', () => {
       .mock({
         name: 'generate_token',
         matcher: (url, opts) =>
-          url === 'https://oauth.me/' && opts.body._streams[7] === 'password',
+          url === 'https://oauth.me/' &&
+          opts.body.get('grant_type') === 'password',
         response: {
           body: {
             access_token: 'an_access_token',
@@ -805,7 +811,7 @@ describe('Fix bugs', () => {
         name: 'refresh_token',
         matcher: (url, opts) =>
           url === 'https://oauth.me/' &&
-          opts.body._streams[1].match('refresh_token'),
+          opts.body.get('refresh_token') === 'refresh_token',
         response: {
           body: {
             access_token: 'a_refreshed_token',
@@ -942,12 +948,12 @@ describe('Test unit of work', () => {
     const cart = await repo.find('/v12/carts/1');
     cart.status = 'foo';
 
-    let updatedCart = await repo.update(cart);
+    await repo.update(cart);
     expect(findLastOptions('put_cart').body).toEqual(
       JSON.stringify({ status: 'foo' })
     );
 
-    updatedCart = await repo.update(cart);
+    await repo.update(cart);
     expect(findLastOptions('put_cart').body).toEqual('{}');
   });
 
@@ -987,7 +993,7 @@ describe('Test unit of work', () => {
       });
 
     const repo = unitOfWorkSdk.getRepository('carts');
-    const cart = await repo.find('/v12/carts/1');
+    await repo.find('/v12/carts/1');
 
     await repo.update({ '@id': '/v12/carts/1', status: null });
     expect(findLastOptions('put_cart').body).toEqual(
@@ -1245,7 +1251,7 @@ describe('Test unit of work', () => {
     };
 
     const repo = unitOfWorkSdk.getRepository('carts');
-    const cart = await repo.create(cartToPost);
+    await repo.create(cartToPost);
 
     expect(fetchMock.lastOptions().body).toEqual(
       JSON.stringify({ order: { '@id': '/v12/orders/1', status: 'waiting' } })
@@ -1333,7 +1339,7 @@ describe('Test unit of work', () => {
 
       cartToPut.set('status', 'refunded');
 
-      const cart = await unitOfWorkSdk.getRepository('carts').update(cartToPut);
+      await unitOfWorkSdk.getRepository('carts').update(cartToPut);
 
       expect(fetchMock.lastOptions().body).toEqual(
         JSON.stringify({