+++ src/TokenStorage.ts
@@ -1,12 +1,30 @@
 /* eslint-disable camelcase */
-import TokenGeneratorInterface from './TokenGenerator/TokenGeneratorInterface';
-import { Token } from './TokenGenerator/types';
+import TokenGeneratorInterface, {
+  TokenBodyReturn,
+} from './TokenGenerator/TokenGeneratorInterface';
+import {
+  ErrorBody,
+  Token,
+  TokenBody,
+  TokenResponse,
+} from './TokenGenerator/types';
 import AsyncStorageInterface from './AsyncStorageInterface';
 import { memoizePromise } from './decorator';
+// eslint-disable-next-line import/no-duplicates
 import type TokenStorageInterface from './TokenStorageInterface';
+// eslint-disable-next-line import/no-duplicates
+import type { HasExpiresAt } from './TokenStorageInterface';
+import {
+  getHttpErrorFromResponse,
+  InvalidGrantError,
+  InvalidScopeError,
+  isOauthError,
+  OauthError,
+} from './ErrorFactory';
 
-interface HasExpiresAt {
-  expires_at: null | number;
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+function isResponse(response: any): response is Response {
+  return typeof response?.json === 'function';
 }
 
 class TokenStorage<T extends Token> implements TokenStorageInterface<T> {
@@ -106,9 +124,11 @@ class TokenStorage<T extends Token> implements TokenStorageInterface<T> {
 
     return this.#tokenGenerator
       .generateToken(parameters)
-      .then((responseData) => {
+      .then(async (response) => {
+        const body = await this.handleTokenResponse(response);
+
         const updatedResponseData = this._addExpiresAtToResponseData(
-          responseData,
+          body,
           callTimestamp
         );
 
@@ -124,9 +144,11 @@ class TokenStorage<T extends Token> implements TokenStorageInterface<T> {
       const callTimestamp = Date.now();
       return this.#tokenGenerator
         .refreshToken(token ? JSON.parse(token) : null)
-        .then((responseData) => {
+        .then(async (response) => {
+          const body = await this.handleTokenResponse(response);
+
           const updatedResponseData = this._addExpiresAtToResponseData(
-            responseData,
+            body,
             callTimestamp
           );
           return this._storeAccessToken(updatedResponseData).then(
@@ -134,6 +156,41 @@ class TokenStorage<T extends Token> implements TokenStorageInterface<T> {
           );
         });
     });
+  }
+
+  private async handleTokenResponse(
+    responseOrBody: TokenBodyReturn<T> | TokenResponse<T>
+  ): Promise<T> {
+    let body: TokenBody<T>;
+    if (isResponse(responseOrBody)) {
+      const response = responseOrBody;
+      try {
+        body = await response.json();
+
+        if (isOauthError(body)) {
+          // throw error if response body is an oauth error
+          this._manageOauthError(body, response);
+        } else if (response.status >= 400) {
+          // throw an error if response status code is an "error" status code
+          throw getHttpErrorFromResponse(response);
+        }
+      } catch (err) {
+        if (!(err instanceof OauthError)) {
+          throw new OauthError(err.type, getHttpErrorFromResponse(response));
+        }
+
+        throw err;
+      }
+    } else {
+      body = responseOrBody;
+
+      if (isOauthError(body)) {
+        // throw error if response body is an oauth error
+        this._manageOauthError(body);
+      }
+    }
+
+    return body;
   }
 
   /**
@@ -163,6 +220,22 @@ class TokenStorage<T extends Token> implements TokenStorageInterface<T> {
       this.accessTokenKey,
       JSON.stringify(responseData)
     );
+  }
+
+  private _manageOauthError(body: ErrorBody, response?: Response): never {
+    let error;
+    if (response) {
+      error = getHttpErrorFromResponse(response);
+    }
+
+    if (body.error === 'invalid_grant') {
+      throw new InvalidGrantError(body.error, error);
+    }
+    if (body.error === 'invalid_scope') {
+      throw new InvalidScopeError(body.error, error);
+    }
+
+    throw new OauthError(body.error, error);
   }
 }