+++ src/TokenStorageInterface.ts
@@ -1,4 +1,4 @@
-interface HasExpiresAt {
+export interface HasExpiresAt {
   // eslint-disable-next-line camelcase
   expires_at: null | number;
 }