+++ src/index.ts
@@ -14,6 +14,7 @@ import {
 import AbstractClient from './client/AbstractClient';
 import TokenStorage from './TokenStorage';
 import AbstractTokenGenerator from './TokenGenerator/AbstractTokenGenerator';
+import AuthorizationCodeFlowTokenGenerator from './TokenGenerator/AuthorizationCodeFlowTokenGenerator';
 import ClientCredentialsGenerator from './TokenGenerator/ClientCredentialsGenerator';
 import PasswordGenerator from './TokenGenerator/PasswordGenerator';
 import ProvidedTokenGenerator from './TokenGenerator/ProvidedTokenGenerator';
@@ -23,14 +24,22 @@ import Mapping from './Mapping';
 import ClassMetadata from './Mapping/ClassMetadata';
 import Attribute from './Mapping/Attribute';
 import Relation from './Mapping/Relation';
-import type { Token } from './TokenGenerator/types';
+import type {
+  Token,
+  ErrorBody,
+  TokenBody,
+  TokenResponse,
+} from './TokenGenerator/types';
 
 import type { SdkMetadata, MetadataDefinition } from './RestClientSdk';
 // eslint-disable-next-line import/no-duplicates
 import type RestClientSdkInterface from './RestClientSdkInterface';
 // eslint-disable-next-line import/no-duplicates
 import type { Config } from './RestClientSdkInterface';
+// eslint-disable-next-line import/no-duplicates
 import type TokenStorageInterface from './TokenStorageInterface';
+// eslint-disable-next-line import/no-duplicates
+import type { HasExpiresAt } from './TokenStorageInterface';
 import type TokenGeneratorInterface from './TokenGenerator/TokenGeneratorInterface';
 import type AsyncStorageInterface from './AsyncStorageInterface';
 
@@ -38,6 +47,7 @@ export default RestClientSdk;
 export {
   AbstractClient,
   AbstractTokenGenerator,
+  AuthorizationCodeFlowTokenGenerator,
   ClientCredentialsGenerator,
   PasswordGenerator,
   ProvidedTokenGenerator,
@@ -61,6 +71,10 @@ export {
 export type {
   SerializerInterface,
   Token,
+  ErrorBody,
+  TokenResponse,
+  TokenBody,
+  HasExpiresAt,
   Config,
   SdkMetadata,
   RestClientSdkInterface,